Remove pb dependency
