+++ ydb/library/kafka/kafka_messages_int.h
@@ -25,23 +25,8 @@ struct TSizeCollector {
     ui32 NumTaggedFields = 0;
 };
 
-template<class T, typename U = std::make_unsigned_t<T>>
-size_t SizeOfUnsignedVarint(T v) {
-    static constexpr U Mask = Max<U>() - 0x7F;
-
-    U value = v;
-    size_t bytes = 1;
-    while ((value & Mask) != 0L) {
-        bytes += 1;
-        value >>= 7;
-    }
-    return bytes;
-}
-
-template<class T>
-size_t SizeOfVarint(T value) {
-    return SizeOfUnsignedVarint(AsUnsigned<T>(value));
-}
+// SizeOfUnsignedVarint/SizeOfVarint now live in kafka.h (next to the varint write logic) and are
+// visible here via kafka_records.h.
 
 template<TKafkaVersion min, TKafkaVersion max>
 constexpr bool VersionAll() {
